+++ aws-crt/native/src/api.h
@@ -6,11 +6,6 @@
  */
 
 #include <aws/common/common.h>
-#include <aws/io/io.h>
-#include <aws/compression/compression.h>
-#include <aws/http/http.h>
-#include <aws/cal/cal.h>
-#include <aws/auth/auth.h>
 
 /* AWS_CRT_API marks a function as public */
 #if defined(_WIN32)
@@ -30,11 +25,14 @@ struct aws_crt_test_struct;
 /* Public function definitions */
 AWS_EXTERN_C_BEGIN
 
+/* CRT */
 AWS_CRT_API void aws_crt_init(void);
-AWS_CRT_API struct aws_crt_event_loop_group *aws_crt_event_loop_group_new(int max_threads);
-AWS_CRT_API void aws_crt_event_loop_group_destroy(struct aws_crt_event_loop_group *crt_elg);
 AWS_CRT_API int aws_crt_test_error(void);
 AWS_CRT_API struct aws_crt_test_struct *aws_crt_test_pointer_error(void);
+
+/* IO */
+AWS_CRT_API struct aws_crt_event_loop_group *aws_crt_event_loop_group_new(uint16_t max_threads);
+AWS_CRT_API void aws_crt_event_loop_group_destroy(struct aws_crt_event_loop_group *crt_elg);
 
 AWS_EXTERN_C_END
